--- src/helpers/errorWithStack.js
@@ -1,9 +1,0 @@
-// @flow
-export default class ErrorWithStack extends Error {
-  constructor(message: ?string, callsite: Function) {
-    super(message);
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, callsite);
-    }
-  }
-}
