+++ src/helpers/errors.js
@@ -1,0 +1,16 @@
+// @flow
+export class ErrorWithStack extends Error {
+  constructor(message: ?string, callsite: Function) {
+    super(message);
+    if (Error.captureStackTrace) {
+      Error.captureStackTrace(this, callsite);
+    }
+  }
+}
+
+export const createLibraryNotSupportedError = (error: Error) =>
+  new Error(
+    `Currently the only supported library to search by text is "react-native".\n\n${
+      error.message
+    }`
+  );
